refactor(router-core): split loadMatches into its own file
